requirements.txt
pillow
pyngrok
